+++ shared/src/iosMain/kotlin/eu/heha/conifer/IosConiferApp.kt
@@ -7,6 +7,7 @@ object IosConiferApp {
         ConiferApp.initialize(
             isDebug = isDebug,
             platform = IosPlatform,
+            dateTimeFormats = IosDateTimeFormats(),
             databaseInitializer = IosDatabaseInitializer,
             syncPrefsInitializer = IosSyncPrefsInitializer,
             credentialsInitializer = IosCredentialsInitializer,